DDCNLS-6515: Moved MetricsService concrete object to wiring.
