+++ app/uk/gov/hmrc/nisp/services/MetricsService.scala
@@ -16,49 +16,8 @@
 
 package uk.gov.hmrc.nisp.services
 
-import com.codahale.metrics.Timer.Context
 import com.codahale.metrics.{Counter, Timer}
-import uk.gov.hmrc.nisp.models.enums.APIType
 import uk.gov.hmrc.nisp.models.enums.APIType.APIType
-import uk.gov.hmrc.play.graphite.MicroserviceMetrics
-
-object MetricsService extends MetricsService with MicroserviceMetrics {
-
-  val timers = Map(
-    APIType.SP -> metrics.defaultRegistry.timer("sp-response-timer"),
-    APIType.NI -> metrics.defaultRegistry.timer("ni-response-timer"),
-    APIType.SchemeMembership -> metrics.defaultRegistry.timer("scheme-membership-response-timer"),
-    APIType.StatePension -> metrics.defaultRegistry.timer("state-pension-response-timer"),
-    APIType.NationalInsurance -> metrics.defaultRegistry.timer("national-insurance-response-timer")
-  )
-
-  val failedCounters = Map(
-    APIType.SP -> metrics.defaultRegistry.counter("sp-failed-counter"),
-    APIType.NI -> metrics.defaultRegistry.counter("ni-failed-counter"),
-    APIType.SchemeMembership -> metrics.defaultRegistry.counter("scheme-membership-failed-counter"),
-    APIType.StatePension -> metrics.defaultRegistry.counter("state-pension-failed-counter"),
-    APIType.NationalInsurance -> metrics.defaultRegistry.counter("national-insurance-failed-counter")
-  )
-
-  override val keystoreReadTimer = metrics.defaultRegistry.timer("keystore-read-timer")
-  override val keystoreWriteTimer = metrics.defaultRegistry.timer("keystore-write-timer")
-
-  override val keystoreReadFailed = metrics.defaultRegistry.counter("keystore-read-failed-counter")
-  override val keystoreWriteFailed = metrics.defaultRegistry.counter("keystore-write-failed-counter")
-
-  override val keystoreHitCounter = metrics.defaultRegistry.counter("keystore-hit-counter")
-  override val keystoreMissCounter = metrics.defaultRegistry.counter("keystore-miss-counter")
-
-  override val identityVerificationTimer = metrics.defaultRegistry.timer("identity-verification-timer")
-  override val identityVerificationFailedCounter = metrics.defaultRegistry.counter("identity-verification-failed-counter")
-
-  override val citizenDetailsTimer = metrics.defaultRegistry.timer("citizen-details-timer")
-  override val citizenDetailsFailedCounter = metrics.defaultRegistry.counter("citizen-details-failed-counter")
-
-  override def startTimer(api: APIType): Context = timers(api).time()
-
-  override def incrementFailedCounter(api: APIType): Unit = failedCounters(api).inc()
-}
 
 trait MetricsService {
   def startTimer(api: APIType): Timer.Context
